msm: Synchronize CPU frequency on thread migration and sync this driver with CAF version
Sometimes when foreground threads, crucial for UI, migrate
from a core at higher frequency to the one running at a lower
frequency, it could badly affect performance for cases like
scrolling.
This change tries to raise the frequency of the destination core
to the same value as the source core, if it is not already running
at the same or higher frequency.  In order for the ondemand
governor to have a fair idea of load in the next sample, it also
reschedules the next sample one sampling interval from the
frequency transition after which the ondemand governor again
takes over.

Signed-off-by: Rohit Gupta <rohgup@codeaurora.org>
Change-Id: I9d2fbb166b0cc042d3fd57cea75a5066df4c794b
Signed-off-by: Paul Keith <javelinanddart@gmail.com>

Conflicts:
	drivers/cpufreq/cpufreq_ondemand.c

Conflicts:
	drivers/cpufreq/cpufreq_ondemand.c